requirements.txt
notebook==6.5.4
numpy==1.24.3
matplotlib==3.7.1
imageio==2.28.1
pandas==2.0.1
papermill=2.4.0
scikit-image==0.20.0
pillow==9.5.0
scipy==1.10.1